# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Absensi",
        "description": "API Aplikasi absensi berbasis web (e-absensi) online untuk memantau kinerja karyawan jarak jauh, dan pekerja work from home (WFH).",
        "termsOfService": "http://127.0.0.1:8000/terms/",
        "contact": {
            "email": "bandungcoders@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/post": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get all data post",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/create": {
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "Add a new post to the store",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title of post",
                                        "type": "string"
                                    },
                                    "body": {
                                        "description": "body of post",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/update/{id}": {
            "put": {
                "tags": [
                    "post"
                ],
                "summary": "Update an existing post",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "title of post",
                                        "type": "string"
                                    },
                                    "body": {
                                        "description": "body of post",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/post/delete/{id}": {
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "Deletes a post",
                "description": "",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/user/delete/{id}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Deletes a user by id",
                "description": "",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}